pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Standalone RAG microservice"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.135,<0.136",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.12",
  "pydantic-settings>=2.7",
  "boto3>=1.34",
  "qdrant-client>=1.17.0,<2.0",
  "nats-py>=2.14",
  "llama-index-core>=0.14.0,<0.15",
  "llama-index-llms-ollama",
  "llama-index-embeddings-fastembed",
  "llama-index-readers-file",
  "llama-index-vector-stores-qdrant",
  "streamlit>=1.55.0",
  "requests>=2.32.5",
  "fastembed>=0.7.4",
  "flagembedding>=1.3.0,<2.0",
  "transformers<4.45",
  "llama-index-postprocessor-flag-embedding-reranker>=0.5.0,<0.6",
  "python-docx>=1.1.0,<2.0",
  "python-pptx>=1.0.0,<2.0",
  "pdf2image>=1.17.0,<2.0",
  "faster-whisper>=1.2.1",
  "kokoro>=0.9.4",
  "soundfile>=0.13.1",
  "numpy>=2.4.3",
  "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "testcontainers>=4",
  "ruff>=0.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

 
